refactor CellOdpHeader
