+++ webapp/components/Assessment/DataTable/Table/CellOdpHeader/CellOdpHeader.tsx
@@ -1,27 +1,31 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-import { FRA } from '@core/assessment'
-
+import { FRA, TableDatumODP } from '@core/assessment'
 import * as BasePaths from '@webapp/main/basePaths'
-import Icon from '@webapp/components/icon'
-import Tooltip from '@webapp/components/tooltip'
 import { useCountryIso, useI18n, usePrintView, useUserInfo } from '@webapp/components/hooks'
 
+import Icon from '@webapp/components/icon'
+import Tooltip from '@webapp/components/tooltip'
+
 type Props = {
-  datum: any
+  datum: TableDatumODP
   sectionName: string
 }
-const CellOdpHeader = (props: Props) => {
+
+const CellOdpHeader: React.FC<Props> = (props) => {
+  const { datum, sectionName } = props
+
   const countryIso = useCountryIso()
   const i18n = useI18n()
   const userInfo = useUserInfo()
   const [printView] = usePrintView()
-  const { datum, sectionName } = props
+
   const { name, year, type, draft, odpId } = datum
   const label = name || year
   const odp = type === 'odp'
   const className = odp && !printView ? 'odp-header-cell' : 'fra-table__header-cell'
+
   return (
     <th className={className}>
       {odp && !printView ? (
@@ -37,4 +41,5 @@ const CellOdpHeader = (props: Props) => {
     </th>
   )
 }
+
 export default CellOdpHeader
