+++ webapp/components/Assessment/DataTable/Table/CellOdpHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CellOdpHeader'
