+++ webapp/components/Assessment/DataTable/Table/Row/rowData/cell/calculated.ts
